CORE-19493 Moving FlowStatusCleanupProcessor and FlowStatusDeletionExecutor from /processors to /components
