+++ components/flow/flow-rest-resource-service-impl/src/test/kotlin/net/corda/flow/rest/impl/FlowStatusDeletionExecutorTest.kt
@@ -1,4 +1,4 @@
-package net.corda.processors.rest
+package net.corda.flow.rest.impl
 
 import net.corda.data.rest.ExecuteFlowStatusCleanup
 import net.corda.data.rest.FlowStatusRecord
